added collsion for end point (which allows the user to start again or quit the game); fixed sharing texture issue; added a floor

+++ SkyboxMazeRunner/SkyboxMazeRunner/MazeMaker.cpp
@@ -34,7 +34,7 @@ bool MazeMaker::LoadMazeFromFile(std::string filename)
 		}
 		else
 		{
-			for (int i = 0; i < line.size(); ++i)
+			for (size_t i = 0; i < line.size(); ++i)
 			{
 				current = line[i] - 48;
 				switch (current)
